Use an in-memory pipe for tests
Prior to this change, the test harness was creating a FIFO on disk using
`mkfifo`. This is really unnecessary, and we can use `pipe` instead to
create a pair of file descriptors corresponding to a FIFO in memory,
without having to go to disk at all.

This change creates the FIFO and sets it to non-blocking to emulate the
existing behavior.